feat: asset mgmt base class
* feat: created AssetManagedPair base class and StablePair inherits from it
* feat: refactored pairs to align interface and implementation
* test: refactor for all asset management tests in one file, avoid duplicates
* fix: removed unused imports
* fix: PR comments
* lib: update forge-std
* feat: parameterize tests for diff curves
* test: added assertions to test for syncing before burning
* fix(interface): removed adjustManagement from IAssetManager
* fix residual merge issues
* test: update gas snapshot
* test: change legacy test timeout value

+++ test/__mocks/AssetManager.sol
@@ -11,7 +11,7 @@ contract AssetManager is IAssetManager
 {
     mapping(IAssetManagedPair => mapping(address => uint112)) public getBalance;
 
-    function adjustManagement(IAssetManagedPair aPair, int224 aToken0Amount, int224 aToken1Amount) external
+    function adjustManagement(IAssetManagedPair aPair, int256 aToken0Amount, int256 aToken1Amount) external
     {
         require(aToken0Amount != type(int224).min && aToken1Amount != type(int224).min, "AM: OVERFLOW");
